requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
requests
google-generativeai
reportlab
